Fix tests on stable/nightly
