+++ benches/distributions/mod.rs
@@ -1,0 +1,3 @@
+mod exponential;
+mod normal;
+mod gamma;
