+++ src/chacha.rs
@@ -203,7 +203,6 @@ impl Rand for ChaChaRng {
 
 #[cfg(test)]
 mod test {
-    use std::iter::order;
     use {Rng, SeedableRng};
     use super::ChaChaRng;
 
@@ -212,8 +211,8 @@ mod test {
         let s = ::test::rng().gen_iter::<u32>().take(8).collect::<Vec<u32>>();
         let mut ra: ChaChaRng = SeedableRng::from_seed(&s[..]);
         let mut rb: ChaChaRng = SeedableRng::from_seed(&s[..]);
-        assert!(order::equals(ra.gen_ascii_chars().take(100),
-                              rb.gen_ascii_chars().take(100)));
+        assert!(::test::iter_eq(ra.gen_ascii_chars().take(100),
+                                rb.gen_ascii_chars().take(100)));
     }
 
     #[test]
@@ -221,8 +220,8 @@ mod test {
         let seed : &[_] = &[0,1,2,3,4,5,6,7];
         let mut ra: ChaChaRng = SeedableRng::from_seed(seed);
         let mut rb: ChaChaRng = SeedableRng::from_seed(seed);
-        assert!(order::equals(ra.gen_ascii_chars().take(100),
-                              rb.gen_ascii_chars().take(100)));
+        assert!(::test::iter_eq(ra.gen_ascii_chars().take(100),
+                                rb.gen_ascii_chars().take(100)));
     }
 
     #[test]
