+++ src/distributions/exponential.rs
@@ -114,26 +114,3 @@ mod test {
         Exp::new(-10.0);
     }
 }
-
-#[cfg(test)]
-mod bench {
-    extern crate test;
-
-    use self::test::Bencher;
-    use std::mem::size_of;
-    use super::Exp;
-    use distributions::Sample;
-
-    #[bench]
-    fn rand_exp(b: &mut Bencher) {
-        let mut rng = ::test::weak_rng();
-        let mut exp = Exp::new(2.71828 * 3.14159);
-
-        b.iter(|| {
-            for _ in 0..::RAND_BENCH_N {
-                exp.sample(&mut rng);
-            }
-        });
-        b.bytes = size_of::<f64>() as u64 * ::RAND_BENCH_N;
-    }
-}
