+++ src/distributions/gamma.rs
@@ -381,39 +381,3 @@ mod test {
         }
     }
 }
-
-#[cfg(test)]
-mod bench {
-    extern crate test;
-    use self::test::Bencher;
-    use std::mem::size_of;
-    use distributions::IndependentSample;
-    use super::Gamma;
-
-
-    #[bench]
-    fn bench_gamma_large_shape(b: &mut Bencher) {
-        let gamma = Gamma::new(10., 1.0);
-        let mut rng = ::test::weak_rng();
-
-        b.iter(|| {
-            for _ in 0..::RAND_BENCH_N {
-                gamma.ind_sample(&mut rng);
-            }
-        });
-        b.bytes = size_of::<f64>() as u64 * ::RAND_BENCH_N;
-    }
-
-    #[bench]
-    fn bench_gamma_small_shape(b: &mut Bencher) {
-        let gamma = Gamma::new(0.1, 1.0);
-        let mut rng = ::test::weak_rng();
-
-        b.iter(|| {
-            for _ in 0..::RAND_BENCH_N {
-                gamma.ind_sample(&mut rng);
-            }
-        });
-        b.bytes = size_of::<f64>() as u64 * ::RAND_BENCH_N;
-    }
-}
