+++ src/distributions/normal.rs
@@ -190,25 +190,3 @@ mod tests {
         LogNormal::new(10.0, -1.0);
     }
 }
-
-#[cfg(test)]
-mod bench {
-    extern crate test;
-    use self::test::Bencher;
-    use std::mem::size_of;
-    use distributions::{Sample};
-    use super::Normal;
-
-    #[bench]
-    fn rand_normal(b: &mut Bencher) {
-        let mut rng = ::test::weak_rng();
-        let mut normal = Normal::new(-2.71828, 3.14159);
-
-        b.iter(|| {
-            for _ in 0..::RAND_BENCH_N {
-                normal.sample(&mut rng);
-            }
-        });
-        b.bytes = size_of::<f64>() as u64 * ::RAND_BENCH_N;
-    }
-}
