+++ src/reseeding.rs
@@ -147,7 +147,7 @@ impl Default for ReseedWithDefault {
 #[cfg(test)]
 mod test {
     use std::default::Default;
-    use std::iter::{order, repeat};
+    use std::iter::repeat;
     use super::{ReseedingRng, ReseedWithDefault};
     use {SeedableRng, Rng};
 
@@ -192,8 +192,8 @@ mod test {
     fn test_rng_seeded() {
         let mut ra: MyRng = SeedableRng::from_seed((ReseedWithDefault, 2));
         let mut rb: MyRng = SeedableRng::from_seed((ReseedWithDefault, 2));
-        assert!(order::equals(ra.gen_ascii_chars().take(100),
-                              rb.gen_ascii_chars().take(100)));
+        assert!(::test::iter_eq(ra.gen_ascii_chars().take(100),
+                                rb.gen_ascii_chars().take(100)));
     }
 
     #[test]
